Revert "update"
This reverts commit 77373750ff50a955392488da48153d5fc06f0634, reversing
changes made to b99e7a7d384876b92d7e62072d7dbad51d832c7f.

+++ Czolgi/tank.h
@@ -40,7 +40,7 @@ protected:
     bool is_loading = false;
     bool is_rotating = false;
 
-    double rotate_angle = -1000;
+    int rotate_angle = -1000;
 
     QTimer *timer_reload = new QTimer(this);
 };
